examples/flows/chat/chat-with-pdf/requirements.txt
langchain
PyPDF2
faiss-cpu
openai
jinja2
python-dotenv
tiktoken
--extra-index-url https://azuremlsdktestpypi.azureedge.net/promptflow/
promptflow
promptflow-tools